# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SMSE Backend API",
    "description": "RESTful API for the SMSE project backend.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/search": {
      "post": {
        "summary": "Multipart Search Files",
        "description": "Search for files using multipart queries. Supports text queries, single/multiple file uploads, or a combination of both. The system computes embeddings for each part and combines them using arithmetic mean for semantic similarity search.",
        "operationId": "searchFiles",
        "tags": [
          "Search"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Maximum number of results to return"
          },
          {
            "name": "modalities",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "text",
                  "image",
                  "audio"
                ]
              }
            },
            "description": "Filter results by specific modalities",
            "style": "form",
            "explode": true
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "Text query for semantic search",
                    "example": "find images of cats"
                  }
                },
                "additionalProperties": false
              },
              "examples": {
                "textOnly": {
                  "summary": "Text-only search",
                  "value": {
                    "query": "find images of cats"
                  }
                }
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "Optional text query to combine with file uploads",
                    "example": "find similar content"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "One or more files to search similar content for. Supports text, image, and audio files."
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Single file upload (legacy support). Use 'files' for multiple files."
                  }
                },
                "additionalProperties": false
              },
              "examples": {
                "singleFile": {
                  "summary": "Single file search",
                  "description": "Search using a single file"
                },
                "multipleFiles": {
                  "summary": "Multiple files search", 
                  "description": "Search using multiple files of different modalities"
                },
                "textAndFiles": {
                  "summary": "Combined text and files search",
                  "description": "Search using both text query and file uploads"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search completed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Search completed successfully"
                    },
                    "query_id": {
                      "type": "integer",
                      "example": 12345
                    },
                    "query_type": {
                      "type": "string",
                      "enum": [
                        "text",
                        "file", 
                        "multipart"
                      ],
                      "example": "multipart",
                      "description": "Type of query: 'text' for text-only, 'file' for single file, 'multipart' for multiple parts"
                    },
                    "query_parts": {
                      "type": "object",
                      "properties": {
                        "text": {
                          "type": "string",
                          "nullable": true,
                          "example": "find similar content",
                          "description": "Text part of the query if provided"
                        },
                        "files": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "nullable": true,
                          "example": ["image1.jpg", "audio.mp3"],
                          "description": "List of uploaded file names if provided"
                        },
                        "total_parts": {
                          "type": "integer",
                          "example": 3,
                          "description": "Total number of query parts (text + files)"
                        }
                      }
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "content_id": {
                            "type": "integer",
                            "example": 6789
                          },
                          "content_path": {
                            "type": "string",
                            "example": "user123/uploads/document.pdf"
                          },
                          "content_tag": {
                            "type": "boolean",
                            "example": true
                          },
                          "similarity_score": {
                            "type": "number",
                            "format": "float",
                            "example": 0.8947,
                            "description": "Cosine similarity score between query and content"
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "limit": {
                          "type": "integer",
                          "example": 10
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "textSearch": {
                    "summary": "Text-only search result",
                    "value": {
                      "message": "Search completed successfully",
                      "query_id": 12345,
                      "query_type": "text",
                      "query_parts": {
                        "text": "find images of cats",
                        "files": null,
                        "total_parts": 1
                      },
                      "results": [
                        {
                          "content_id": 101,
                          "content_path": "user1/uploads/cat_image.jpg",
                          "content_tag": true,
                          "similarity_score": 0.92
                        }
                      ],
                      "pagination": {
                        "limit": 10
                      }
                    }
                  },
                  "multipartSearch": {
                    "summary": "Multipart search result",
                    "value": {
                      "message": "Search completed successfully", 
                      "query_id": 12346,
                      "query_type": "multipart",
                      "query_parts": {
                        "text": "similar visual content",
                        "files": ["reference.jpg", "audio_sample.wav"],
                        "total_parts": 3
                      },
                      "results": [
                        {
                          "content_id": 202,
                          "content_path": "user1/uploads/similar_image.png",
                          "content_tag": true,
                          "similarity_score": 0.87
                        }
                      ],
                      "pagination": {
                        "limit": 10
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. No valid query parts provided or unsupported file type.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "examples": [
                        "No valid query parts provided. Either query text or files are required",
                        "Unsupported file type '.xyz' in file 'test.xyz'. Allowed types: .txt, .pdf, .jpg, .png, .mp3, .wav"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error. Error processing query or creating embeddings.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "examples": [
                        "Error creating embedding for query",
                        "Error processing query: SMSE service unavailable",
                        "Error performing search: Database connection failed"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Query History",
        "description": "Retrieve the search query history for the authenticated user.",
        "operationId": "getQueryHistory",
        "tags": [
          "Search"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Maximum number of queries to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of queries to skip (for pagination)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved query history.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "queries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 123
                          },
                          "text": {
                            "type": "string",
                            "example": "example search query"
                          },
                          "timestamp": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2025-03-03T12:34:56Z"
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "example": 45
                        },
                        "limit": {
                          "type": "integer",
                          "example": 10
                        },
                        "offset": {
                          "type": "integer",
                          "example": 0
                        },
                        "has_more": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/search/{query_id}": {
      "get": {
        "summary": "Get Search Results History",
        "description": "Retrieve the search results history for a specific query.",
        "operationId": "getSearchResultsHistory",
        "tags": [
          "Search"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the query to retrieve search results history for."
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved search results history.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 123
                        },
                        "text": {
                          "type": "string",
                          "example": "example search query"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2025-03-03T12:34:56Z"
                        }
                      }
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "content_id": {
                            "type": "integer",
                            "example": 456
                          },
                          "similarity_score": {
                            "type": "number",
                            "format": "float",
                            "example": 0.87
                          },
                          "retrieved_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2025-03-03T12:40:00Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Query not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Query not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Query",
        "description": "Deletes a specific query by its ID.",
        "operationId": "deleteQuery",
        "tags": [
          "Search"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the query to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Query deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Query deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Query not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Query not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
}